+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 0.0.3
+
+* Add swift-argument-parser support with help
+* Open xcworkspace as well as xcodeproj
+* Optionally (-all) also open Package.swift
+* Open docs
+* Create new files and open them in Xcode
+* Open with Xcode in background
+
 ## 0.0.2
 
 Roll back Swift tools to 5.1 to support Mojave